+++ Data/__init__.py
@@ -1,2 +1,3 @@
 from .experiment_descriptor import ExperimentDescriptor
 from .user import User
+from .vnf import Vnf