+++ Data/experiment_descriptor.py
@@ -1,4 +1,5 @@
 from .user import User
+from .vnf import Vnf
 from Helper import Serialize
 from typing import Dict, Tuple, List
 
@@ -12,7 +13,7 @@ class ExperimentDescriptor:
 
     @staticmethod
     def validate(data: Dict) -> Tuple[bool, List[str]]:
-        keys = ['Id', 'Name', 'Platform', 'TestCases', 'UEs', 'User']
+        keys = ['Id', 'Name', 'Platform', 'TestCases', 'UEs', 'User', 'Slice', 'NSD']
         return Serialize.CheckKeys(data, *keys)
 
     @property
@@ -34,6 +35,15 @@ class ExperimentDescriptor:
     @property
     def UEs(self) -> Dict[str, Dict]:
         return self._data['UEs']
+
+    @property
+    def HasNsd(self) -> bool:
+        # TODO: Modify handling of NSDs to allow multiple network services
+        return self._data['NSD'] is not None
+
+    @property
+    def Slice(self) -> str:
+        return self._data['Slice']
 
     @property
     def ValidityCheck(self) -> Tuple[bool, List[str]]: