+++ Executor/Tasks/PreRun/add_entry.py
@@ -3,9 +3,9 @@ from Helper import Level
 from time import sleep
 
 
-class AddExperimentEntry(Task):
+class AddExecutionEntry(Task):
     def __init__(self, logMethod):
-        super().__init__("Add Experiment Entry", None, logMethod, None)
+        super().__init__("Add Execution Entry", None, logMethod, None)
 
     def Run(self):
         self.Log(Level.INFO, 'Sending entry information')