+++ Executor/executor_base.py
@@ -31,6 +31,14 @@ class ExecutorBase(Child):
         self.AddMessage("Init")
 
     @property
+    def Params(self) -> Dict:
+        return self.params
+
+    @property
+    def HasFailed(self) -> bool:
+        return self.hasFailed
+
+    @property
     def Descriptor(self) -> Optional[ExperimentDescriptor]:
         return self.params.get('Descriptor', None)