+++ Executor/post_runner.py
@@ -1,5 +1,5 @@
 from typing import Dict
-from .Tasks.PostRun import RequestResults, SaveResults, UpdateExperimentEntry
+from .Tasks.PostRun import RequestResults, SaveResults, UpdateExecutionEntry, Decommission
 from .executor_base import ExecutorBase
 from tempfile import TemporaryDirectory
 
@@ -11,11 +11,14 @@ class PostRunner(ExecutorBase):
     def Run(self):
         self.SetStarted()
 
+        Decommission(self.Log, self.Configuration.PostRunParams).Start()
+        self.AddMessage('Resources decommisioned', 10)
+
         RequestResults(self.Log).Start()
         self.AddMessage('Results received', 30)
         SaveResults(self.Log).Start()
         self.AddMessage('Results stored', 60)
-        UpdateExperimentEntry(self.Log).Start()
+        UpdateExecutionEntry(self.Log).Start()
         self.AddMessage('Entry updated', 90)
 
         self.SetFinished(percent=100)