+++ Helper/__init__.py
@@ -1,7 +1,9 @@
-from .log import Log, Level, LogInfo
+from .log import Log, LogInfo
+from .log_level import Level
 from .config import Config, TapConfig
 from .child import Child
 from .serialize import Serialize
 from .tap_executor import Tap
 from .cli_executor import Cli
 from .dashboard_generator import DashboardGenerator
+from .influx import InfluxDb, InfluxPayload, InfluxPoint