+++ Helper/log_level.py
@@ -1,0 +1,6 @@
+from enum import Enum, unique
+
+
+@unique
+class Level(Enum):
+    DEBUG, INFO, WARNING, ERROR, CRITICAL = range(5)