+++ Interfaces/__init__.py
@@ -1,2 +1,3 @@
 from .management import Management
 from .dispatcher import DispatcherApi
+from .portal import PortalApi